# swchallenge1/Presentation/Dtos/UpdateTaskRequest.cs
namespace swchallenge1.Presentation.Dtos;

public class UpdateTaskRequest
{
    public string? Title { get; set; }
    public string? Description { get; set; }
    public bool IsComplete { get; set; }
}